// src/locator/models/element-context.model.ts
export interface ElementContext {
    "data-testid"?: string;
    "data-test"?: string;
    "data-qa"?: string;
    id?: string;
    name?: string;
    nameAttribute?: string;
    ariaLabel?: string;
    "aria-label"?: string;
    role?: string;
    placeholder?: string;
    label?: string;
    title?: string;
    alt?: string;
    class?: string;
    href?: string;
    type?: string;
    disabled?: boolean;
    visible?: boolean;
    required?: boolean;
    readonly?: boolean;
    text?: string;
    tagName?: string;
    cssPath?: string;
    relativeCssPath?: string;
    xpath?: string;
    relativeXpath?: string;
}